requirements.txt
annotated-types==0.7.0
anyio==4.7.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
filelock==3.16.1
Flask==3.1.0
fsspec==2024.12.0
googlesearch-python==1.2.5
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.2.1
packaging==24.2
pdfminer==20191125
pdfminer.six==20240706
pillow==11.0.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
safetensors==0.4.5
setuptools==75.6.0
sniffio==1.3.1
soupsieve==2.6
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
